Place options into a separete file
